requirements.txt
# DermAI Backend Requirements
# Install with: pip install -r requirements.txt

# Flask Framework
flask==3.0.0
flask-cors==4.0.0

# Deep Learning
tensorflow==2.15.0
# For CPU only, use: tensorflow-cpu==2.15.0

# Image Processing
Pillow==10.1.0

# Scientific Computing
numpy==1.24.3

# Optional: For production deployment
gunicorn==21.2.0

# Optional: For enhanced API features
python-dotenv==1.0.0
